Safari-611.1.10/LayoutTests/fast/xpath/nsresolver-exception-expected.txt

CONSOLE MESSAGE: Error: this is an exception
SUCCESS
